+++ poutyne/framework/metrics/__init__.py
@@ -20,7 +20,6 @@ You should have received a copy of the GNU Lesser General Public License along w
 # pylint: disable=wildcard-import
 from .metrics_registering import *
 from .pytorch_loss_functions_registering import *
-from .epoch_metrics import *
 from .decomposable import *
 from .predefined import *
 from .utils import *