+++ tests/framework/metrics/test_metrics_model_integration.py
@@ -30,7 +30,7 @@ import torch
 import torch.nn as nn
 import torchmetrics
 
-from poutyne import Model, EpochMetric, rename_doubles, do_register_metric_class, unregister_metric_class
+from poutyne import Model, Metric, rename_doubles, do_register_metric_class, unregister_metric_class
 
 
 class MyConstTorchMetric(torchmetrics.Metric):
@@ -47,15 +47,18 @@ class MyConstTorchMetric(torchmetrics.Metric):
         return self.value
 
 
-class ConstEpochMetric(EpochMetric):
+class ConstMetric(Metric):
     def __init__(self, value=0):
         super().__init__()
         self.value = value
 
-    def forward(self, y_pred, y_true):
+    def forward(self, y_pred: torch.Tensor, y_true: torch.Tensor):
         return self.value
 
-    def get_metric(self):
+    def update(self, y_pred: torch.Tensor, y_true: torch.Tensor):
+        pass
+
+    def compute(self):
         return self.value
 
     def reset(self):
@@ -69,8 +72,8 @@ def get_batch_metric(value):
     return some_metric_name
 
 
-class SomeMetricName(ConstEpochMetric):
-    def get_metric(self):
+class SomeMetricName(ConstMetric):
+    def compute(self):
         return torch.FloatTensor([self.value])
 
     def reset(self):
@@ -147,7 +150,7 @@ class MetricsModelIntegrationTest(unittest.TestCase):
 
     def test_batch_metrics_with_multiple_names_returned_by_dict(self):
         d = dict(zip(self.metric_names, self.metric_values))
-        batch_metric = ConstEpochMetric(d)
+        batch_metric = ConstMetric(d)
         model = Model(
             self.pytorch_network, self.optimizer, self.loss_function, batch_metrics=[(self.metric_names, batch_metric)]
         )
@@ -155,7 +158,7 @@ class MetricsModelIntegrationTest(unittest.TestCase):
 
     def test_epoch_metrics_with_multiple_names_returned_by_dict(self):
         d = dict(zip(self.metric_names, self.metric_values))
-        epoch_metric = ConstEpochMetric(d)
+        epoch_metric = ConstMetric(d)
         model = Model(
             self.pytorch_network, self.optimizer, self.loss_function, epoch_metrics=[(self.metric_names, epoch_metric)]
         )
@@ -169,17 +172,17 @@ class MetricsModelIntegrationTest(unittest.TestCase):
         self._test_history(model, self.metric_names, self.metric_values)
 
     def test_epoch_metrics_with_multiple_names_returned_by_tensor(self):
-        epoch_metric = ConstEpochMetric(torch.tensor(self.metric_values))
+        epoch_metric = ConstMetric(torch.tensor(self.metric_values))
         model = Model(
             self.pytorch_network, self.optimizer, self.loss_function, epoch_metrics=[(self.metric_names, epoch_metric)]
         )
         self._test_history(model, self.metric_names, self.metric_values)
 
     def test_epoch_metrics_with_name_with_multiple_names_returned_by_tensor(self):
-        class EpochMetricWithName(ConstEpochMetric):
+        class MetricWithName(ConstMetric):
             __name__ = self.metric_names
 
-        epoch_metric = EpochMetricWithName(torch.tensor(self.metric_values))
+        epoch_metric = MetricWithName(torch.tensor(self.metric_values))
         model = Model(self.pytorch_network, self.optimizer, self.loss_function, epoch_metrics=[epoch_metric])
         self._test_history(model, self.metric_names, self.metric_values)
 
@@ -199,7 +202,7 @@ class MetricsModelIntegrationTest(unittest.TestCase):
     @skipIf(not torch.cuda.is_available(), "no gpu available")
     def test_epoch_metrics_with_multiple_names_returned_by_tensor_on_gpu(self):
         with torch.cuda.device(MetricsModelIntegrationTest.cuda_device):
-            epoch_metric = ConstEpochMetric(torch.tensor(self.metric_values).cuda())
+            epoch_metric = ConstMetric(torch.tensor(self.metric_values).cuda())
             model = Model(
                 self.pytorch_network,
                 self.optimizer,
@@ -217,7 +220,7 @@ class MetricsModelIntegrationTest(unittest.TestCase):
         self._test_history(model, self.metric_names, self.metric_values)
 
     def test_epoch_metrics_with_multiple_names_returned_by_ndarray(self):
-        epoch_metric = ConstEpochMetric(np.array(self.metric_values))
+        epoch_metric = ConstMetric(np.array(self.metric_values))
         model = Model(
             self.pytorch_network, self.optimizer, self.loss_function, epoch_metrics=[(self.metric_names, epoch_metric)]
         )
@@ -231,7 +234,7 @@ class MetricsModelIntegrationTest(unittest.TestCase):
         self._test_history(model, self.metric_names, self.metric_values)
 
     def test_epoch_metrics_with_multiple_names_returned_by_list(self):
-        epoch_metric = ConstEpochMetric(list(self.metric_values))
+        epoch_metric = ConstMetric(list(self.metric_values))
         model = Model(
             self.pytorch_network, self.optimizer, self.loss_function, epoch_metrics=[(self.metric_names, epoch_metric)]
         )
@@ -245,7 +248,7 @@ class MetricsModelIntegrationTest(unittest.TestCase):
         self._test_history(model, self.metric_names, self.metric_values)
 
     def test_epoch_metrics_with_multiple_names_returned_by_tuple(self):
-        epoch_metric = ConstEpochMetric(tuple(self.metric_values))
+        epoch_metric = ConstMetric(tuple(self.metric_values))
         model = Model(
             self.pytorch_network, self.optimizer, self.loss_function, epoch_metrics=[(self.metric_names, epoch_metric)]
         )
@@ -341,30 +344,6 @@ class MetricsModelIntegrationTest(unittest.TestCase):
         valid_y = torch.randint(10, (dataset_size,))
 
         model = Model(nn.Linear(1, 10), 'sgd', 'cross_entropy', epoch_metrics=['f1', ('f1_2', 'f1')])
-        names = ['fscore_macro', 'f1_2']
-        history = model.fit(
-            train_x,
-            train_y,
-            validation_data=(valid_x, valid_y),
-            batch_size=MetricsModelIntegrationTest.batch_size,
-            epochs=MetricsModelIntegrationTest.epochs,
-        )
-        for logs in history:
-            for name in names:
-                self.assertIn(name, logs)
-                self.assertEqual(ANY, logs[name])
-                self.assertIn('val_' + name, logs)
-                self.assertEqual(ANY, logs['val_' + name])
-
-    def test_torch_metrics_with_str_str_tuple(self):
-        dataset_size = MetricsModelIntegrationTest.batch_size * MetricsModelIntegrationTest.steps_per_epoch
-        torch.manual_seed(42)
-        train_x = torch.rand(dataset_size, 1)
-        train_y = torch.randint(10, (dataset_size,))
-        valid_x = torch.rand(dataset_size, 1)
-        valid_y = torch.randint(10, (dataset_size,))
-
-        model = Model(nn.Linear(1, 10), 'sgd', 'cross_entropy', torch_metrics=['f1', ('f1_2', 'f1')])
         names = ['fscore_macro', 'f1_2']
         history = model.fit(
             train_x,